Добавляет ДЗ#5
