+++ js/functions.js
@@ -1,21 +1,6 @@
-// function checkStringLength
+const checkStringLength = (str, maxLength) => str.length <= maxLength;
 
-function checkStringLength (str, maxLength) {
-
-  return str.length <= maxLength;
-}
-
-// function isPalindrome
-
-function isPalindrome (str) {
-  const reverseStr = (str.split('').reverse().join(''));
-
-  return str === reverseStr;
-}
-
-// function extractNumbers
-
-function extractNumbers (str) {
+const extractNumbers = (str) => {
   let numberStr = '';
 
   for (let i = 0; i <= str.length; i++) {
@@ -30,44 +15,8 @@ function extractNumbers (str) {
   }
 
   return NaN;
-}
+};
 
-// function Addletter
+checkStringLength('function', 5);
 
-function formingStringChar (str, num, addStr) {
-  const start = str.length;
-  const finish = addStr.length;
-  const count = num - start;
-  let newStr = '';
-  let secStr = '';
-
-  if (str.length >= num) {
-    return str;
-  }
-
-  if (finish === 1) {
-    return addStr.repeat(count) + str;
-  }
-
-  if (finish > count) {
-    for (let i = 0; i < count; i++) {
-      newStr += addStr[i];
-    }
-
-    return newStr + str;
-  }
-
-  for (const char of addStr) {
-    newStr += char;
-  }
-  for (let i = 0; i < count - finish; i++) {
-    secStr += addStr[i];
-  }
-
-  return secStr + newStr + str;
-}
-
-checkStringLength ('function',5);
-isPalindrome ('система');
-extractNumbers ('а2 я4 т4омат');
-formingStringChar ('q', 7, 'wes');
+extractNumbers('local2 56gg');
